% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{as.OP.lp_problem}
\alias{as.OP.lp_problem}
\title{Create a \code{\link[ROI:OP]{ROI::OP()}} Object.}
\usage{
\method{as.OP}{lp_problem}(x)
}
\arguments{
\item{x}{An \code{\link[=lp_problem]{lp_problem()}}.}
}
\value{
An \code{OP} object as returned from \code{\link[ROI:OP]{ROI::OP()}}.
}
\description{
Convert an \code{\link[=lp_problem]{lp_problem()}} object to a \code{\link[ROI:OP]{ROI::OP()}} object.
Used internally in \code{\link[=lp_solve]{lp_solve()}}.
}
\examples{
library(ROI.plugin.highs)

problem <- lp_problem() |>
    lp_var(x[a = 1:2, b = 1:2], upper = 10) |>
    lp_alias(sum_x = sum(x)) |>
    lp_max(sum_x + 5)

# Solve directly with lp_solve()
direct_solution <- lp_solve(problem)
direct_solution

# Or run each step manually
op <- as.OP(problem)
solution_raw <- ROI_solve(op)
solution_pretty <- pretty_solution(problem, solution = solution_raw)

# pretty_solution() gives each variable its dim and dimnames.
solution_raw$solution
solution_pretty$variables

# It also computes aliases
solution_pretty$aliases

# Notice that in solution_raw, the objective value is 40 instead of 45.
# The objective function is (sum(x) + 5), but the addend (+5) is irrelevant
# for finding the optimum, so ROI ignores it.
solution_raw$objval
solution_pretty$objective
}
\seealso{
\code{\link[=pretty_solution]{pretty_solution()}} to prettify the solution returned by \code{\link[ROI:ROI_solve]{ROI::ROI_solve()}}.
}
