% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{bind_vars}
\alias{bind_vars}
\title{Concatenate variables and numbers.}
\usage{
bind_vars(...)
}
\arguments{
\item{...}{Problem variables and/or numeric constants, vectors or arrays.}
}
\value{
An \code{lp_variable}.
}
\description{
Bind multiple variables and/or constants together.
}
