% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{lp_constraint}
\alias{lp_constraint}
\alias{lp_con}
\alias{lp_subject_to}
\title{Add Constraints to an \code{\link[=lp_problem]{lp_problem()}}}
\usage{
lp_constraint(.problem, ...)

lp_con(.problem, ...)

lp_subject_to(.problem, ...)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{...}{One or more linear constraints. Can be named. They must:
\itemize{
\item Contain one or more variables defined with \code{\link[=lp_variable]{lp_variable()}}
\item Contain a comparison operator, such as \code{<=},  \code{==} or \verb{=>}.
\item If it's a nonlinear constraint, it must be written as:

\code{\link[=nonlinear]{nonlinear()}}\verb{ <= number}
}}
}
\value{
The \code{.problem} with added \verb{$constraints}. Previous constraints are not
overwritten, so it's possible to call \code{\link[=lp_constraint]{lp_constraint()}} multiple times without
overwriting previously defined constraints.

The \verb{$constraints} inherit from \code{\link[ROI:L_constraint]{ROI::L_constraint()}}, \code{\link[ROI:Q_constraint]{ROI::Q_constraint()}}
or \code{\link[ROI:F_constraint]{ROI::F_constraint()}}.
\itemize{
\item Quadratic constraints are represented as

\eqn{\frac{1}{2} x'Q_{i}x + L_{i}x \le \text{rhs}_{i} \qquad \forall i}
\item While nonlinear constraints are represented as

\eqn{F(x)_i \le \text{rhs}_i \qquad \forall i}
}
}
\description{
Restrict the variables in an \code{\link[=lp_problem]{lp_problem()}} with linear or quadratic constraints.
}
\examples{
# Order Constraint ------------------------------
## ordered[i] > ordered[i-1]  forall i in 2:n
n <- 4
p <- lp_problem() |>
    lp_variable(ordered[1:n], lower = 0)

## Three alternatives
pc <- p |> lp_constraint(
    alt1 = ordered[2:n] > ordered[1:(n-1)],
    alt2 = for (i in 2:n) ordered[i] > ordered[i-1]
)

## The 'for' construct can be wrapped around 'lp_constraint' instead of inside it
for (i in 2:n) {
    pc <- pc |> lp_constraint(
        alt3 = ordered[i] > ordered[i-1]
    )
}

## The only difference are the row names of the constraint matrix when printing
print(pc$constraints)


# Nonlinear Constraint --------------------------
## log(x + 1) > y  =>  log(x + 1) - y > 0
nlp <- lp_problem() |> 
    lp_variable(x, lower = 0) |> 
    lp_variable(y) |> 
    lp_constraint(
        # Nonlinear constraints must always be written as `nonlinear(...) <= number`
        nl_con = nonlinear(log(x + 1) - y) > 0
    )

print(nlp$constraints)
}
