% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{lp_delete_constraint}
\alias{lp_delete_constraint}
\title{Delete constraints}
\usage{
lp_delete_constraint(.problem, names)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{names}{Characted vector with the names of constraints to be deleted.
It is not possible to delete unnamed constraints, so make sure to name them
if you plan to delete them later.}
}
\description{
Remove named constraints from an \code{\link[=lp_problem]{lp_problem()}}.
}
