% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{lp_fix_vars}
\alias{lp_fix_vars}
\title{Fix Variables to a Value}
\usage{
lp_fix_vars(.problem, ...)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{...}{Name-value pairs, variables with their respective values.
\code{NA} values are not fixed, and will remain free within their bounds.}
}
\value{
The \code{.problem} with modified lower and upper bounds for the variables.
}
\description{
Set the lower and upper bounds of variables to a fixed value.
}
