% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{lp_minimize}
\alias{lp_minimize}
\alias{lp_maximize}
\alias{lp_min}
\alias{lp_max}
\title{Set an Objective Function}
\usage{
lp_minimize(.problem, objective)

lp_maximize(.problem, objective)

lp_min(.problem, objective)

lp_max(.problem, objective)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{objective}{Expression to optimize. Can be:
\itemize{
\item The number 0, in which case the solver will attempt to find any feasible solution.
\code{\link[=lp_find_feasible]{lp_find_feasible()}} serves the same purpose.
\item A linear or quadratic expression containing decision variables.
\item A nonlinear expression wrapped in \code{\link[=nonlinear]{nonlinear()}}.
}}
}
\value{
The \code{.problem} with the new \verb{$objective} function.

The \verb{$objective} inherits from \code{\link[ROI:L_objective]{ROI::L_objective()}}, \code{\link[ROI:Q_objective]{ROI::Q_objective()}},
or \code{\link[ROI:F_objective]{ROI::F_objective()}}.
\itemize{
\item A quadratic objective function is represented as

\eqn{\frac{1}{2} x'Qx + Lx}
\item While a nonlinear objective function is simply represented as

\eqn{F(x)}
}
}
\description{
Minimize of maximize a linear or quadratic expression.
}
\details{
If \code{objective} evaluates to a multivariate variable instead of a scalar, it will
apply \code{sum(objective)} and display a message. Suppress this message by writing
the \code{sum} yourself.
}
\examples{
# Linear Objective using an Alias ---------------
profit   <- c(Phone = 60, Tablet = 20, eBook = 10)
max_made <- c(Phone = 500, Tablet = 300, eBook = 950)
product  <- names(profit)
fix_cost <- 11e3

p <- lp_problem() |>
    lp_variable(made[product], lower = 0, upper = max_made, integer = TRUE) |>
    lp_alias(total_profit = sum(made * profit)) |>
    lp_maximize(total_profit - fix_cost) |>
    lp_constraint(sum(made) <= 1500)

p$objective

library(ROI.plugin.highs)
s <- lp_solve(p)
print(s)

s$aliases$total_profit
sum(c(p$objective$L) * s$variables_vec)

s$aliases$total_profit - fix_cost
s$objective


# Nonlinear objective ---------------------------
nlp <- lp_problem() |> 
    lp_variable(x, lower = 0) |> 
    lp_variable(y, lower = 0) |> 
    lp_maximize(nonlinear(sqrt(x) * log(y))) |> 
    lp_constraint(x + y <= 10)

# There are some different solvers within `nloptr`
library(ROI.plugin.nloptr)
lpsugar_applicable_solvers(nlp)

lp_solve(
    nlp, 
    solver = "nloptr.cobyla", 
    start = list(x = 1, y = 1)
)

# See more examples in the Nonlinear vignette
# vignette("nonlinear", package = "lpsugar")
}
\seealso{
\code{\link[=nonlinear]{nonlinear()}} For general nonlinear optimization.
}
