% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{lp_solve}
\alias{lp_solve}
\alias{lp_find_feasible}
\title{Solve a Linear Problem}
\usage{
lp_solve(.problem, solver, ..., start, binary_as_logical = FALSE)

lp_find_feasible(.problem, binary_as_logical = FALSE, ...)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{solver}{String specifying the solver to use.
If missing, then the default solver returned by \code{\link[ROI:ROI_options]{ROI::ROI_options()}} is used.}

\item{...}{Control arguments to be passed on to the solver.}

\item{start}{Start value of variables, for nonlinear solvers. One of:
\itemize{
\item Named list of variables with their respective values.
If a variable is missing, it is set to \code{pmax(0, lower)}.
\item An \code{lp_solution} object as returned by \code{\link[=lp_solve]{lp_solve()}} or \code{\link[=lp_find_feasible]{lp_find_feasible()}}.
\item A vector containing the values of each variable, one after another.
}}

\item{binary_as_logical}{Boolean. If \code{FALSE} (the default), binary variables
are returned as \verb{\{0, 1\}}. If \code{TRUE}, binary variables are returned as logical \verb{\{FALSE, TRUE\}}.}
}
\value{
A list with the following fields:
\itemize{
\item \verb{$objective} : Scalar, value of the objective function at optimum.
\item \verb{$variables} : List of arrays, values of the variables at optimum.
\item \verb{$aliases} : List of arrays, values of the aliases at optimum.
\item \verb{$variables_vec} : Numeric vector, values of the variables at optimum.
\item \verb{$status} : Status as returned by \code{\link[ROI:ROI_solve]{ROI::ROI_solve()}}.
\item \verb{$message} : Message as returned by \code{\link[ROI:ROI_solve]{ROI::ROI_solve()}},
\item \verb{$op} : Optimization Problem \code{OP}, as returned by \code{\link[ROI:OP]{ROI::as.OP()}}.
}
}
\description{
\code{lp_solve()} computes the optimal solution of an \code{\link[=lp_problem]{lp_problem()}}, whereas
\code{lp_find_feasible()} returns an arbitrary feasible solution.
}
\examples{
library(ROI.plugin.highs)

r <- letters[1:2]
c <- LETTERS[1:3]

# Problem with Optimal Solution
p_opt <- lp_problem() |>
    lp_variable(x[r, c], binary = TRUE) |>
    lp_alias(total = sum(x)) |>
    lp_maximize(total + 1) |>
    lp_constraint(x[1, 1] + x[2, 1] <= 1L)

s1 <- lp_solve(p_opt)
s1$status
s1$objective
s1$aliases$total
s1$variables$x

s2 <- lp_solve(p_opt, binary_as_logical = TRUE)
s2$variables$x


# Problem with Unbounded Solution
p_unb <- lp_problem() |>
    lp_variable(y) |>
    lp_max(y)

s3 <- lp_solve(p_unb)
s3$status
s3$objective # Might change depending on solver

# Infeasible Problem
p_no <- lp_problem() |>
    lp_variable(z[letters]) |>
    lp_constraint(z[1] <= z[1] - 10)

s5 <- lp_find_feasible(p_no)

s5$status
s5$variables$z
s5$objective
}
\seealso{
\code{\link[=as.OP.lp_problem]{as.OP.lp_problem()}}, \code{\link[=pretty_solution]{pretty_solution()}}.
}
