% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsugar-package.R
\docType{package}
\name{lpsugar-package}
\alias{lpsugar}
\alias{lpsugar-package}
\title{lpsugar: Linear, Quadratic and Nonlinear Programming in R}
\description{
Algebraic Modeling Language for Optimization in R. Supports linear, quadratic, or nonlinear programming. Problems can be solved with a variety of solvers using 'ROI'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benet1one/lpsugar}
  \item \url{https://benet1one.github.io/lpsugar/}
  \item Report bugs at \url{https://github.com/benet1one/lpsugar/issues}
}

}
\author{
\strong{Maintainer}: Bernat Garcia \email{bernatgarciaquadrat@gmail.com} [copyright holder]

}
\keyword{internal}
