% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{lpsugar_applicable_solvers}
\alias{lpsugar_applicable_solvers}
\alias{lpsugar_available_solvers}
\title{List Applicable and Available Solvers}
\usage{
lpsugar_applicable_solvers(problem)

lpsugar_available_solvers(problem)
}
\arguments{
\item{problem}{An \code{\link[=lp_problem]{lp_problem()}} or a \code{\link[ROI:OP]{ROI::OP()}}.}
}
\value{
\itemize{
\item \code{lpsugar_applicable_solvers} returns a character vector with the solver names.
\item \code{lpsugar_available_solvers} returns a \code{data.frame} with information on the solvers.
}
}
\description{
Show which solvers can solve a problem. Applicable solvers must be installed and loaded,
whereas Available solvers needn't be installed.
}
\details{
\itemize{
\item \code{lpsugar_applicable_solvers} returns a character vector of solver names, which
can be used in \verb{lp_solve(solver = _)}. It lists solvers which:
\itemize{
\item Can solve the \code{problem}.
\item Are installed and have been loaded with \code{library(ROI)} or \verb{library(ROI.plugin.<solver>)}.
}
\item \code{lpsugar_available_solvers} returns a \code{data.frame} with information on the solvers.
It lists solvers which:
\itemize{
\item Can solve the \code{problem}.
\item Do not need to be installed.
}
}

Note since Nonlinear Solvers are also applicable to Linear and Quadratic problems,
they will also be listed.
}
\examples{
library(ROI)

quadratic_prob <- lp_problem() |> 
    lp_var(x, lower = 0) |> 
    lp_var(y, lower = 0) |> 
    lp_min(x^2 + y) |> 
    lp_con(x + y >= 10)

# Installed and loaded quadratic solvers
lpsugar_applicable_solvers(quadratic_prob)
# All quadratic solvers
lpsugar_available_solvers(quadratic_prob) [c("Package", "Version", "Repository")]
}
\seealso{
\code{\link[ROI:ROI_applicable_solvers]{ROI::ROI_applicable_solvers()}}, \code{\link[ROI:ROI_available_solvers]{ROI::ROI_available_solvers()}}.
}
