% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{parameter}
\alias{parameter}
\title{Set Dimensions and Names for Parameters}
\usage{
parameter(.x, ..., byrow = TRUE)
}
\arguments{
\item{.x}{Vector or matrix of any type.}

\item{...}{Dimnames. If you want to return a vector, it needs one element;
if you want to return a matrix it needs two elements.}

\item{byrow}{Boolean, true by default. Only used if \code{.x} is a vector and \code{...} has two elements.}
}
\value{
A named vector or matrix.
}
\description{
Return a named vector or matrix.
}
\examples{
my_set <- letters[1:3]
my_parameter <- c(2, 6, 3) |> parameter(my_set)
my_parameter

rows <- letters[1:2]
cols <- LETTERS[1:3]
my_matrix <- c(
    1, 2, 3,
    4, 5, 6
) |> parameter(rows, cols)
my_matrix

# Also works if .x is already a matrix.
mat <- matrix(1:6, nrow = 3, ncol = 2)
mat |> parameter(r = letters[1:3], c = LETTERS[1:2])
}
