% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{pretty_solution}
\alias{pretty_solution}
\title{Prettify the Solution of a Model.}
\usage{
pretty_solution(problem, solution, binary_as_logical = FALSE)
}
\arguments{
\item{problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{solution}{A list as returned by \code{\link[ROI:ROI_solve]{ROI::ROI_solve()}}.}

\item{binary_as_logical}{Boolean. If \code{FALSE} (the default), binary variables
are returned as \verb{\{0, 1\}}. If \code{TRUE}, binary variables are returned as logical \verb{\{FALSE, TRUE\}}.}
}
\value{
A list with the following fields:
\itemize{
\item \verb{$objective} : Scalar, value of the objective function at optimum.
\item \verb{$variables} : List of arrays, values of the variables at optimum.
\item \verb{$aliases} : List of arrays, values of the aliases at optimum.
\item \verb{$variables_vec} : Numeric vector, values of the variables at optimum.
\item \verb{$status} : Status as returned by \code{\link[ROI:ROI_solve]{ROI::ROI_solve()}}.
\item \verb{$message} : Message as returned by \code{\link[ROI:ROI_solve]{ROI::ROI_solve()}},
\item \verb{$op} : Optimization Problem \code{OP}, as returned by \code{\link[ROI:OP]{ROI::as.OP()}}.
}
}
\description{
Takes a problem and its solution and prettifies the solution. Used internally
in \code{\link[=lp_solve]{lp_solve()}}.
}
\examples{
library(ROI.plugin.highs)

problem <- lp_problem() |>
    lp_var(x[a = 1:2, b = 1:2], upper = 10) |>
    lp_alias(sum_x = sum(x)) |>
    lp_max(sum_x + 5)

# Solve directly with lp_solve()
direct_solution <- lp_solve(problem)
direct_solution

# Or run each step manually
op <- as.OP(problem)
solution_raw <- ROI_solve(op)
solution_pretty <- pretty_solution(problem, solution = solution_raw)

# pretty_solution() gives each variable its dim and dimnames.
solution_raw$solution
solution_pretty$variables

# It also computes aliases
solution_pretty$aliases

# Notice that in solution_raw, the objective value is 40 instead of 45.
# The objective function is (sum(x) + 5), but the addend (+5) is irrelevant
# for finding the optimum, so ROI ignores it.
solution_raw$objval
solution_pretty$objective
}
\seealso{
\code{\link[=lp_solve]{lp_solve()}} for the standard way to solve a problem.

\code{\link[=as.OP.lp_problem]{as.OP.lp_problem()}} to convert an \code{lp_problem} to an Optimization Problem \code{(OP)} object
from package \code{ROI}.
}
