% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugarlp-package.R
\docType{package}
\name{sugarlp-package}
\alias{sugarlp}
\alias{sugarlp-package}
\title{sugarlp: Algebraic Modeling Language for Linear Problems}
\description{
Algebraic Modeling Language inspired in R's vectorized syntax. Define and solve Linear Programming Problems by calling 'lpSolveAPI', which supports Mixed Integer Linear Programming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benet1one/sugarlp}
  \item Report bugs at \url{https://github.com/benet1one/sugarlp/issues}
}

}
\author{
\strong{Maintainer}: benet \email{benet1one@gmail.com}

}
\keyword{internal}
