% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fun.R
\name{sum_over}
\alias{sum_over}
\title{Index Based Summation}
\usage{
sum_over(...)
}
\arguments{
\item{...}{The first argument(s) must be named: the name represents the indexing variable,
and the values represent the sequence over which to sum.

The last argument is the expression of the sum.}
}
\value{
The value of the sum.
}
\description{
Sum over one or more indexing variables.
}
\details{
The syntax is similar to the one used in math. For instance,
\eqn{ \sum_{i=1}^n \sum_{j=1}^m {x_{ij} * c_j} }
would be written as
\code{sum_over(i = 1:n, j = 1:n, x[i,j] * c[j])}.
}
\examples{
cost <- c(5, 2, 7)
p <- lp_problem() |>
  lp_variable(x[1:2, 1:3]) |>
  lp_minimize(sum_over(i = 1:2, j = 1:3, x[i, j] * cost[j]))
p$objective
}
