% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{lp_alias}
\alias{lp_alias}
\alias{lp_implicit_variable}
\alias{lp_impvar}
\title{Define an Alias or Implicit Variable (IMPVAR)}
\usage{
lp_alias(.problem, ...)

lp_implicit_variable(.problem, ...)

lp_impvar(.problem, ...)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{...}{}
}
\description{
Define an Alias or Implicit Variable (IMPVAR)
}
