% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{lp_alias_manual}
\alias{lp_alias_manual}
\alias{lp_impvar_manual}
\title{Manually Define an Alias by Assigning its Values}
\usage{
lp_alias_manual(.problem, definition, expression)

lp_impvar_manual(.problem, definition, expression)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}} object.}

\item{definition}{Name and dimensions of the variable.
\itemize{
\item If the variable is a scalar, simply type it's name.
\itemize{
\item \code{lp_variable(x)}
}
\item If the variable is a vector, type it's name and indices. Indices can also be named.
\itemize{
\item \code{lp_variable( v[1:5] )}
\item \code{lp_variable( v[letters[1:5]] )}
\item \code{lp_variable( v[ind = letters[1:5]] )}
\item \verb{ind <- letters[1:5]; lp_variable( v[ind] )}
}

The last two have the same result.
\item If the variable is a matrix or n-dimensional array, type it's name and the indices
for every dimension (\code{\link[base:dimnames]{base::dimnames()}}), comma-separated.
\itemize{
\item \code{lp_variable( mat[1:2, 1:3] )}
\item \code{lp_variable( arr[1:2, 1:3, 1:2] )}
}
}}

\item{expression}{Code to assign values to the alias. The values must be
numeric or \verb{<lp_variable>}. See examples.}
}
\value{
The \code{.problem} with the added alias in \verb{$aliases}.
}
\description{
This function does the same as \code{\link[=lp_alias]{lp_alias()}}, but gives more freedom.
It provides a better syntax for defining aliases that cannot be defined
in a single line of code.
}
\examples{
# We have a 4x4 matrix X
# We want an alias A, of length 4, such that
# A[i] <- i * X[i, i+1]
# A[4] <- 4

p <- lp_problem() |> 
    lp_variable(
        X[1:4, 1:4]
    ) |> 
    lp_alias_manual(
        A[1:4], {
            for (i in 1:3) A[i] <- i * X[i, i+1]
            A[4] <- 4
        }
    )
}
\seealso{
\code{\link[=lp_alias]{lp_alias()}}
}
