% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{lp_constraint}
\alias{lp_constraint}
\alias{lp_con}
\alias{lp_subject_to}
\title{Add constraints to an \code{\link[=lp_problem]{lp_problem()}}}
\usage{
lp_constraint(.problem, ...)

lp_con(.problem, ...)

lp_subject_to(.problem, ...)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{...}{One or more linear constraints. Can be named. They must:
\itemize{
\item Contain one or more variables defined with \code{\link[=lp_variable]{lp_variable()}}
\item Contain a comparison operator ( \verb{< / <= / == / => / >} )
}}
}
\value{
The \code{.problem} with added \verb{$constraints}. (Note: previous constraints are not
overritten).

Constraints can be represented as \verb{lhs * vars <dir> rhs}.

The \verb{$constraints} field has the following subfields:
\itemize{
\item \verb{$lhs} : \code{\link[slam:matrix]{slam::simple_triplet_matrix()}} where each row is a constraint,
each column is a variable, and the values represent coefficients.
\item \verb{$dir} : Character vector with elements \code{"<="}, \code{"=="}, or \code{">="},
the direction of each constraint.
\item \verb{$rhs} : Numeric column vector representing the right hand side of each constraint.
\item \verb{$name} : Character vector with the names of the constraints, if \code{...} is named,
or \code{""} for unnamed constraints.
\item \verb{$call} : Expression that defined each constraint.
}
}
\description{
Restrict the variables in an \code{\link[=lp_problem]{lp_problem()}} with linear constraints.
}
\examples{
# Ordered variable constraint
## ordered[i] > ordered[i-1]  for all i in 2:n
n <- 4
p <- lp_problem() |>
    lp_variable(ordered[1:n], lower = 0)

## Three alternatives
pc <- p |> lp_constraint(
    alt1 = ordered[2:n] > ordered[1:(n-1)],
    alt2 = for (i in 2:n) ordered[i] > ordered[i-1]
)

## The 'for' construct can be wrapped around 'lp_constraint' instead of inside it
for (i in 2:n) {
    pc <- pc |> lp_constraint(
        alt3 = ordered[i] > ordered[i-1]
    )
}

## The only difference are the rownames of the constraint matrix
print(pc$constraints)
}
