% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{lp_constraint}
\alias{lp_constraint}
\alias{lp_con}
\alias{lp_subject_to}
\title{Add constraints to an \code{\link[=lp_problem]{lp_problem()}}}
\usage{
lp_constraint(.problem, ...)

lp_con(.problem, ...)

lp_subject_to(.problem, ...)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{...}{One or more linear constraints. Can be named. They must:
\itemize{
\item Contain one or more variables defined with \code{\link[=lp_variable]{lp_variable()}}
\item Contain a comparison operator ( \verb{< / <= / == / => / >} )
}}
}
\value{
An \code{\link[=lp_problem]{lp_problem()}} with added \verb{$constraints}. (Note: previous constraints are not
overritten).

Constraints can be represented as \verb{lhs * vars <dir> rhs}.

The \verb{$constraints} field has the following subfields:
\itemize{
\item \verb{$lhs} : Matrix where each row is a constraint, each column is a variable,
and the values represent coefficients.
\item \verb{$dir} : Character vector with elements \code{"<="}, \code{"=="}, or \code{">="},
the direction of each constraint.
\item \verb{$rhs} : Numeric column vector representing the right hand side of each constraint.
\item \verb{$name} : Character vector with the names of the constraints, if \code{...} is named,
or \code{""} for unnamed constraints.
\item \verb{$call} : Expression that defined each constraint.
}
}
\description{
Restrict the variables in an \code{\link[=lp_problem]{lp_problem()}} with linear constraints.
}
