% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{lp_fix_vars}
\alias{lp_fix_vars}
\title{Fix Variables to a Value}
\usage{
lp_fix_vars(.problem, ...)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{...}{Name-value pairs, variables with their respective values.
\code{NA} values are not fixed, and will remain free within their bounds.}
}
\value{
The \code{.problem} with modified lower and upper bounds for the variables.
}
\description{
Set the lower and upper bounds of variables to a fixed value.
}
\examples{
# Variable `x` has 5 values
fixed_x <- numeric(5)
# The first value of `x` should be 1, last value should be 9
fixed_x[1] <- 1
fixed_x[5] <- 9
# Other values are free
fixed_x[2:4] <- NA

p <- lp_problem() |> 
    lp_var(x[1:5], lower = 0) |> 
    lp_fix_vars(x = fixed_x)

rbind(
    lower = p$variables$x$lower,
    upper = p$variables$x$upper
)
}
