% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{lp_minimize}
\alias{lp_minimize}
\alias{lp_maximize}
\title{Set an objective function}
\usage{
lp_minimize(.problem, objective)

lp_maximize(.problem, objective)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{objective}{Expression to optimize, which must evaluate to an \code{lp_variable} object.
Alternatively, set \code{objective = 0} to remove the objective function and let the solver
find a feasible solution instead of optimizing.}
}
\value{
An \code{\link[=lp_problem]{lp_problem()}} object with the updated objective function.
}
\description{
Minimize of maximize an expression.
}
\details{
The objective function can be found in \code{.problem$objective}, a list with the following
information:
\itemize{
\item \verb{$coef} : Vector with the coefficients for each variable.
\item \verb{$add} : Numeric, addend to the final value. It is not used in the solver.
\item \verb{$direction} : Character, goal of the solver. Can be \code{"minimize"} or \code{"maximize"}.
}

If \code{objective} evaluates to a multivariate variable instead of a scalar, it will
apply \code{sum(objective)} and display a message.
}
