% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{lp_minimize}
\alias{lp_minimize}
\alias{lp_maximize}
\alias{lp_min}
\alias{lp_max}
\title{Set an objective function}
\usage{
lp_minimize(.problem, objective)

lp_maximize(.problem, objective)

lp_min(.problem, objective)

lp_max(.problem, objective)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{objective}{Expression to optimize, which must evaluate to an \code{lp_variable} object.
Alternatively, set \code{objective = 0} to remove the objective function and let the solver
find a feasible solution instead of optimizing, like \code{\link[=lp_find_feasible]{lp_find_feasible()}}.}
}
\value{
The \code{.problem} with the updated \verb{$objective} function, a list with these fields:
\itemize{
\item \verb{$coef} : Vector with the coefficients for each variable.
\item \verb{$add} : Numeric, addend to the final value. It is not used in the solver.
\item \verb{$direction} : String, goal of the solver. Can be \code{"minimize"} or \code{"maximize"}.
\item \verb{$expr} : String, expression that defined the objective function.
}
}
\description{
Minimize of maximize an expression.
}
\details{
If \code{objective} evaluates to a multivariate variable instead of a scalar, it will
apply \code{sum(objective)} and display a message. Suppress this message by writing
the \code{sum} yourself.
}
