% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.R
\name{lp_minimize_function}
\alias{lp_minimize_function}
\alias{lp_maximize_function}
\alias{lp_min_fun}
\alias{lp_max_fun}
\title{Set a General Nonlinear Objective Function}
\usage{
lp_minimize_function(.problem, fun, gradient = NULL, hessian = NULL)

lp_maximize_function(.problem, fun, gradient = NULL, hessian = NULL)

lp_min_fun(.problem, fun, gradient = NULL, hessian = NULL)

lp_max_fun(.problem, fun, gradient = NULL, hessian = NULL)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{fun}{Function to optimize. The function's arguments must match all
defined \link[=lp_variable]{variables}.}

\item{gradient}{Function that returns the gradient vector.}

\item{hessian}{Function that returns the hessian matrix.}
}
\value{
An \code{\link[=lp_problem]{lp_problem()}} with the new \verb{$objective} function.
}
\description{
Optimize an R function.
}
\examples{
library(ROI.plugin.alabama)

## Simple Example -----------------------
# max  sqrt(x) * log(y)
#  st  x + y <= 10

p1 <- lp_problem() |> 
    lp_variable(x, lower = 0) |> 
    lp_variable(y, lower = 0) |> 
    lp_maximize_function(\(x, y) sqrt(x) * log(y)) |> 
    lp_constraint(x + y <= 10)

lp_solve(p1, start = list(x = 1, y = 1))
}
\seealso{
\code{\link[=lp_minimize]{lp_minimize()}} to optimize linear or quadratic functions.
}
