% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.R
\name{lp_minimize_function}
\alias{lp_minimize_function}
\alias{lp_maximize_function}
\alias{lp_min_fun}
\alias{lp_max_fun}
\title{Set a General Nonlinear Objective Function}
\usage{
lp_minimize_function(.problem, fun, gradient = NULL, hessian = NULL)

lp_maximize_function(.problem, fun, gradient = NULL, hessian = NULL)

lp_min_fun(.problem, fun, gradient = NULL, hessian = NULL)

lp_max_fun(.problem, fun, gradient = NULL, hessian = NULL)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{fun}{Function to optimize. The function's arguments must match all
defined \link[=lp_variable]{variables}.}

\item{gradient}{Function that returns the gradient vector.}

\item{hessian}{Function that returns the hessian matrix.}
}
\value{
An \code{\link[=lp_problem]{lp_problem()}} with the new \verb{$objective} function.
}
\description{
Optimize an R function.
}
\seealso{
\code{\link[=lp_minimize]{lp_minimize()}} to optimize linear or quadratic functions.
}
