% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp_problem}
\alias{lp_problem}
\title{Create a Linear Problem}
\usage{
lp_problem()
}
\value{
An \code{lp_problem} object with fields:
\itemize{
\item \verb{$variables} : List of \code{\link[=lp_variable]{lp_variable()}}s.
\item \verb{$objective} : List with information about the objective function,
set with \code{\link[=lp_minimize]{lp_minimize()}} or \code{\link[=lp_maximize]{lp_maximize()}}.
\item \verb{$constraints} : List of constraints added with \code{\link[=lp_constraint]{lp_constraint()}}.
}
}
\description{
Use \code{\link[=lp_variable]{lp_variable()}} to define the variables,
\code{\link[=lp_minimize]{lp_minimize()}} or \code{\link[=lp_maximize]{lp_maximize()}} to define the objective function,
\code{\link[=lp_constraint]{lp_constraint()}} to add constraints and
\code{\link[=lp_solve]{lp_solve()}} to find the optimum.
}
\examples{
# ROI needs to be imported for lp_solve() to work
library(ROI)

lp_problem() |>
    lp_variable(x[1:2], lower = 0) |>
    lp_maximize(x[1] + x[2]) |>
    lp_constraint(
        2*x[1] +   x[2] <= 8,
        2*x[1] + 3*x[2] <= 12
    ) |>
    lp_solve()

# For a Quick Start guide see `vignette("lpsugar", package = "lpsugar")`
}
