% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{lp_solve}
\alias{lp_solve}
\title{Solve a Linear Problem}
\usage{
lp_solve(
  .problem,
  binary_as_logical = FALSE,
  unbound_as_inf = TRUE,
  report_status = FALSE,
  ...
)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{binary_as_logical}{Boolean. If \code{FALSE} (the default), binary variables
are returned as \verb{\{0, 1\}}. If \code{TRUE}, binary variables are returned as logical \verb{\{FALSE, TRUE\}}.}

\item{unbound_as_inf}{Boolean, whether to replace very large numbers with \code{+Inf} and
very small numbers with \code{-Inf}.}

\item{report_status}{Boolean, whether to emit a message indicating the status number and
description of the solver. Some common status are:
\itemize{
\item \verb{0 | optimal solution found}
\item \verb{2 | the model is infeasible}
\item \verb{3 | the model is unbounded}
For a full list of status and their meaning see \code{\link[lpSolveAPI:solve.lpExtPtr]{lpSolveAPI::solve.lpExtPtr()}}.
}}

\item{...}{Control parameters passed to \code{\link[lpSolveAPI:lp.control]{lpSolveAPI::lp.control()}}. For a full list of
options see \code{\link[lpSolveAPI:lp.control.options]{lpSolveAPI::lp.control.options()}}.}
}
\description{
Attempt to compute the optimal solution of an \code{\link[=lp_problem]{lp_problem()}}.
}
