% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{lp_solve}
\alias{lp_solve}
\title{Solve a Linear Problem}
\usage{
lp_solve(
  .problem,
  binary_as_logical = FALSE,
  unbound_as_inf = TRUE,
  report_status = FALSE,
  verbose = "severe",
  ...
)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{binary_as_logical}{Boolean. If \code{FALSE} (the default), binary variables
are returned as \verb{\{0, 1\}}. If \code{TRUE}, binary variables are returned as logical \verb{\{FALSE, TRUE\}}.}

\item{unbound_as_inf}{Boolean, whether to replace very large numbers with \code{+Inf} and
very small numbers with \code{-Inf}.}

\item{report_status}{Boolean, whether to emit a message indicating the status number and
description of the solver. For a full list of status and their meaning
see \code{\link[lpSolveAPI:solve.lpExtPtr]{lpSolveAPI::solve.lpExtPtr()}}.Some common status are:
\itemize{
\item \verb{0 | optimal solution found}
\item \verb{2 | the model is infeasible}
\item \verb{3 | the model is unbounded}
}}

\item{verbose}{String indicating the severity of messages reported by \code{lp_solve}.
\itemize{
\item \code{"neutral"} : No reporting.
\item \code{"critical"} : Only critical messages are reported. Hard errors like instability, out of memory, etc.
\item \code{"severe"} :    Only severe messages are reported. Errors.
\item \code{"important"} : Only important messages are reported. Warnings and Errors.
\item \code{"normal"} :    Normal messages are reported.
\item \code{"detailed"} : Detailed messages are reported. Like model size, continuing B&B improvements, etc.
\item \code{"full"} : All messages are reported. Useful for debugging purposes and small models.
}}

\item{...}{Control parameters passed to \code{\link[lpSolveAPI:lp.control]{lpSolveAPI::lp.control()}}. For a full list of
options see \code{\link[lpSolveAPI:lp.control.options]{lpSolveAPI::lp.control.options()}}.}
}
\description{
Attempt to compute the optimal solution of an \code{\link[=lp_problem]{lp_problem()}}.
}
