% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{lp_variable}
\alias{lp_variable}
\alias{lp_var}
\title{Define a variable for an \code{\link[=lp_problem]{lp_problem()}}}
\usage{
lp_variable(
  .problem,
  definition,
  integer = FALSE,
  binary = FALSE,
  lower = -Inf,
  upper = +Inf
)

lp_var(
  .problem,
  definition,
  integer = FALSE,
  binary = FALSE,
  lower = -Inf,
  upper = +Inf
)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}} object.}

\item{definition}{Expression.
\itemize{
\item If the variable is a scalar, simply type it's name.
\itemize{
\item \code{lp_variable(x)}
}
\item If the variable is a vector, type it's name and indices. Indices can also be named.
\itemize{
\item \code{lp_variable( v[1:5] )}
\item \code{lp_variable( v[letters[1:5]] )}
\item \code{lp_variable( v[ind = letters[1:5]] )}
\item \verb{ind <- letters[1:5]; lp_variable( v[ind] )}
}

The last two have the same result.
\item If the variable is a matrix or n-dimensional array, type it's name and the indices
for every dimension (\code{\link[base:dimnames]{base::dimnames()}}), comma-separated.
\itemize{
\item \code{lp_variable( mat[1:2, 1:3] )}
\item \code{lp_variable( arr[1:2, 1:3, 1:2] )}
}
}}

\item{integer}{Boolean, whether to treat variable as integer.}

\item{binary}{Boolean, whether to treat variable as binary, \{0, 1\}.}

\item{lower}{Numeric scalar or array. Lower bound for the variable.}

\item{upper}{Numeric scalar or array. Upper bound for the variable.}
}
\value{
The \code{.problem} with an added variable in \verb{$variables}.
The fields of \code{lp_variable} objects are intended for internal use, modifying them is
highly discouraged.
\itemize{
\item \verb{$name} : String, name of the variable.
\item \verb{$lower} and \verb{$upper} : Bounds.
\item \verb{$type} : String, one of \code{"real"}, \code{"integer"} or \code{"binary"}.
\item \verb{$integer} and \verb{$binary} : Booleans. If \verb{$binary} is true, then \verb{$integer} is also true.
}

The following fields are meant for internal use only.
\itemize{
\item \verb{$ind} : Integer array. Indicates which indices correspond to this variable.
Meant for internal use only.
\item \verb{$coef} : Numeric matrix. The number of rows is the length of the variable, the number
of columns is the total amount of variables in the problem.
The values represent coefficients that are modified when adding variables or multiplying
by a constant. Used for objective, constraints and aliases.
\item \verb{$add} : Numeric column vector. The number of rows is the length of the variable.
The values represent addends, modified when adding the variable and a constant.
Used for objective, constraints and aliases.
\item \verb{$raw} : Boolean indicating if the variable has been modified in any way
(indexed, multiplied, summed, ...) or remains as defined.
}
}
\description{
Add a variable to a problem by defining it's name, dimensions, type, and bounds.
}
\examples{

# Diet problem
## You can only eat as much as you have of each food.

pantry <- c(Cookies = 8, Bread = 5, Milk = 1, Cereal = 3)
food <- names(pantry)

lp_problem() |>
    lp_variable(diet[food], lower = 0, upper = pantry)


# Graph problem
## active[i,j] = 1  if the edge from i to j is active.

nodes <- paste("node", 1:10)

lp_problem() |>
    lp_variable(active[nodes, nodes], binary = TRUE)

}
