% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{lp_variable}
\alias{lp_variable}
\title{Define a variable for an \code{\link[=lp_problem]{lp_problem()}}}
\usage{
lp_variable(
  .problem,
  definition,
  integer = FALSE,
  binary = FALSE,
  lower = -Inf,
  upper = +Inf
)
}
\arguments{
\item{.problem}{An \code{\link[=lp_problem]{lp_problem()}} object.}

\item{definition}{Expression.
\itemize{
\item If the variable is a scalar, simply type it's name.
\itemize{
\item \code{lp_variable(x)}
}
\item If the variable is a vector, type it's name and indices. Indices can also be named.
\itemize{
\item \code{lp_variable( v[1:5] )}
\item \code{lp_variable( v[letters[1:5]] )}
\item \code{lp_variable( v[ind = letters[1:5]] )}
\item \verb{ind <- letters[1:5]; lp_variable( v[ind] )}
}

The last two have the same result.
\item If the variable is a matrix or n-dimensional array, type it's name and the indices
for every dimension (\code{\link[base:dimnames]{base::dimnames()}}), comma-separated.
\itemize{
\item \code{lp_variable( mat[1:2, 1:3] )}
\item \code{lp_variable( arr[1:2, 1:3, 1:2] )}
}
}}

\item{integer}{Boolean, whether to treat variable as integer.}

\item{binary}{Boolean, whether to treat variable as binary, \{0, 1\}.}

\item{lower}{Numeric scalar. Lower bound for the variable.}

\item{upper}{Numeric scalar. Upper bound for the variable.}
}
\value{
An \code{\link[=lp_problem]{lp_problem()}} object with an added variable in \verb{$variables}.
}
\description{
Add a variable to a problem by defining it's name, dimensions, type, and bounds.
}
\details{
Bounds must be numeric scalars. If you want to bind each index of the variable separately,
use \code{\link[=lp_constraint]{lp_constraint()}}.
}
