% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsugar-package.R
\docType{package}
\name{lpsugar-package}
\alias{lpsugar}
\alias{lpsugar-package}
\title{lpsugar: Algebraic Modeling Language for Linear Problems}
\description{
Algebraic Modeling Language inspired in R's vectorized syntax. Define Mixed Integer Linear Programming (MILP) problems and solve them with 'ROI'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benet1one/lpsugar}
  \item Report bugs at \url{https://github.com/benet1one/lpsugar/issues}
}

}
\author{
\strong{Maintainer}: benet \email{benet1one@gmail.com}

}
\keyword{internal}
