% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{make_model}
\alias{make_model}
\title{Make a Linear Program Model}
\usage{
make_model(problem, verbose = "severe", ...)
}
\arguments{
\item{problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{verbose}{String indicating the severity of messages reported by \code{lp_solve}.
\itemize{
\item \code{"neutral"} : No reporting.
\item \code{"critical"} : Only critical messages are reported. Hard errors like instability, out of memory, etc.
\item \code{"severe"} :    Only severe messages are reported. Errors.
\item \code{"important"} : Only important messages are reported. Warnings and Errors.
\item \code{"normal"} :    Normal messages are reported.
\item \code{"detailed"} : Detailed messages are reported. Like model size, continuing B&B improvements, etc.
\item \code{"full"} : All messages are reported. Useful for debugging purposes and small models.
}}

\item{...}{Control parameters passed to \code{\link[lpSolveAPI:lp.control]{lpSolveAPI::lp.control()}}. For a full list of
options see \code{\link[lpSolveAPI:lp.control.options]{lpSolveAPI::lp.control.options()}}.}
}
\value{
An \code{lpExtPtr} pointer from the \code{lpSolveAPI} package.
}
\description{
Create an \code{lpExtPtr} pointer from the \code{lpSolveAPI} package.
Used internally in \code{\link[=lp_solve]{lp_solve()}}.
}
\examples{
p <- lp_problem() |>
    lp_var(x[a = 1:2, b = 1:2], upper = 10) |>
    lp_max(sum(x) + 5)

# Solve directly with lp_solve()
direct_solution <- lp_solve(p)
direct_solution

# Or do each step
model <- make_model(p)
solution_raw <- solve_model(model)
solution_pretty <- pretty_solution(problem = p, solution = solution_raw)

# pretty_solution() gives each variable its dim and dimnames.
solution_raw$variables
solution_pretty$variables

# Notice that in solution_raw, the objective value is 40 instead of 45.
# The objective function is (sum(x) + 5), but the addend (+5) is irrelevant
# for finding the optimum, so lpSolveAPI ignores it.
solution_raw$objective
solution_pretty$objective
}
