% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.R
\name{nonlinear}
\alias{nonlinear}
\alias{nl}
\title{Nonlinear Expression}
\usage{
nonlinear(expr)

nl(expr)
}
\arguments{
\item{expr}{(Unquoted) expression containing
decision variables. Must return a numeric vector
when evaluated.}
}
\value{
The quoted expression with class
\code{"nonlinear"}
}
\description{
Nonlinear transformations and operations on variables.
}
\examples{
# max  sqrt(x) * log(y)
#  st  x^y <= 10

p <- lp_problem() |> 
    lp_variable(x, lower = 0) |> 
    lp_variable(y, lower = 1) |> 
    lp_maximize(nonlinear(sqrt(x) * log(y))) |> 
    lp_constraint(nonlinear(x^y) <= 5)

print(p)

library(ROI.plugin.nloptr)
s <- lp_solve(p, solver = "nloptr.cobyla", start = list(x = 2, y = 2))
print(s)
}
