% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear.R
\name{nonlinear}
\alias{nonlinear}
\title{Nonlinear Expression}
\usage{
nonlinear(expr)
}
\arguments{
\item{expr}{(Unquoted) expression containing
decision variables. Must return a numeric vector
when evaluated.}
}
\value{
The quoted expression with class
\code{"nonlinear_lp_variable"}
}
\description{
Nonlinear transformations and operations on variables.
}
\examples{
library(ROI.plugin.nloptr)

## Simple Example -----------------------
# max  sqrt(x) * log(y)
#  st  x + y <= 10

p1 <- lp_problem() |> 
    lp_variable(x, lower = 0) |> 
    lp_variable(y, lower = 0) |> 
    lp_maximize(nonlinear(sqrt(x) * log(y))) |> 
    lp_constraint(x + y <= 10)

# There are some different solvers within `nloptr`
lpsugar_applicable_solvers(p1)
lp_solve(p1, solver = "nloptr.cobyla", start = list(x = 1, y = 1))

# See more examples in the Nonlinear vignette
# vignette("nonlinear", package = "lpsugar")
}
