% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{solution_summary}
\alias{solution_summary}
\alias{constraint_summary}
\title{Compute a Summary of a Solution or Point}
\usage{
solution_summary(problem, solution, tol = 2e-06)

constraint_summary(problem, solution, tol = 2e-06)
}
\arguments{
\item{problem}{An \code{\link[=lp_problem]{lp_problem()}}.}

\item{solution}{One of:
\itemize{
\item Named list of variables with their respective values.
If a variable is missing, it is set to \code{max(0, lower)}.
\item An \code{lp_solution} object as returned by \code{\link[=lp_solve]{lp_solve()}}.
\item A vector containing the values of each variable, one after another.
}}

\item{tol}{Tolerance to use for checking constraint and bound feasibility}
}
\value{
A named list with the computed statistics.
}
\description{
Check feasibility, constraint saturation, calculate objective value and aliases.
}
