% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_model}
\alias{solve_model}
\title{Solve a Model}
\usage{
solve_model(model)
}
\arguments{
\item{model}{A pointer of class \code{lpExtPtr} created with \code{\link[=make_model]{make_model()}} or \code{\link[lpSolveAPI:make.lp]{lpSolveAPI::make.lp()}}.}
}
\description{
Find the optimal solution of a model created with \code{\link[=make_model]{make_model()}} or \code{\link[lpSolveAPI:make.lp]{lpSolveAPI::make.lp()}}.
Used internally in \code{\link[=lp_solve]{lp_solve()}}.
}
\section{Value}{

A list with fields:
\itemize{
\item \verb{$objective} : Numeric scalar, value of the objective function at the optimum.
NOTE: This value does not include the addend \code{(problem$objective$add)}.
\item \verb{$variables} : Numeric vector with the value of the variables at the optimum.
\item \verb{$status_number} : Integer indicating the status of the solver.
\item \verb{$status_description} : String describing the status of the solver.
\item \verb{$pointer} : Pointer to the \code{lpSolveAPI} model, class \code{lpExtPtr}.
}
}

