% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_model}
\alias{solve_model}
\title{Solve a Model}
\usage{
solve_model(model)
}
\arguments{
\item{model}{A pointer of class \code{lpExtPtr} created with \code{\link[=make_model]{make_model()}} or \code{\link[lpSolveAPI:make.lp]{lpSolveAPI::make.lp()}}.}
}
\description{
Find the optimal solution of a model created with \code{\link[=make_model]{make_model()}} or \code{\link[lpSolveAPI:make.lp]{lpSolveAPI::make.lp()}}.
Used internally in \code{\link[=lp_solve]{lp_solve()}}.
}
\section{Value}{

A list with fields:
\itemize{
\item \verb{$objective} : Numeric scalar, value of the objective function at the optimum.
NOTE: This value does not include the addend \code{(problem$objective$add)}.
\item \verb{$variables} : Numeric vector with the value of the variables at the optimum.
\item \verb{$status_number} : Integer indicating the status of the solver.
\item \verb{$status_description} : String describing the status of the solver.
\item \verb{$pointer} : Pointer to the \code{lpSolveAPI} model, class \code{lpExtPtr}.
}
}

\examples{
p <- lp_problem() |>
    lp_var(x[a = 1:2, b = 1:2], upper = 10) |>
    lp_max(sum(x) + 5)

# Solve directly with lp_solve()
direct_solution <- lp_solve(p)
direct_solution

# Or do each step
model <- make_model(p)
solution_raw <- solve_model(model)
solution_pretty <- pretty_solution(problem = p, solution = solution_raw)

# pretty_solution() gives each variable its dim and dimnames.
solution_raw$variables
solution_pretty$variables

# Notice that in solution_raw, the objective value is 40 instead of 45.
# The objective function is (sum(x) + 5), but the addend (+5) is irrelevant
# for finding the optimum, so lpSolveAPI ignores it.
solution_raw$objective
solution_pretty$objective
}
